sockets: Move common socket code to separate module.
This code is effectively a layer below the TCTI. The resourcemgr
uses it internally but exposing it through the socket TCTI interface
isn't appropriate. Moving this code out to a separate module will allow
us to break this tight coupling between the resroucemgr and that TCTI
more easily.
