+++ resourcemgr/resourceMgr.vcxproj
@@ -88,6 +88,7 @@
     <ClInclude Include="..\..\include\Implementation.h" />
     <ClInclude Include="..\common\debug.h" />
     <ClInclude Include="..\common\platform.h" />
+    <ClInclude Include="..\common\sockets.h" />
     <ClInclude Include="..\common\syscontext.h" />
     <ClInclude Include="..\include\tcti\tcti_socket.h" />
     <ClInclude Include="..\sysapi\include\tpm20.h" />
@@ -103,6 +104,7 @@
   <ItemGroup>
     <ClCompile Include="..\common\debug.c" />
     <ClCompile Include="..\common\getcommands.c" />
+    <ClCompile Include="..\common\sockets.cpp" />
     <ClCompile Include="..\common\syscontext.c" />
     <ClCompile Include="..\tcti\commonchecks.c" />
     <ClCompile Include="..\tcti\platformcommand.c" />
